add confirmation dialog to Goals.vue & Goal.vue
delete todos attached to a goal when the goal is deleted

+++ src/components/Goals.vue
@@ -40,6 +40,21 @@
           <v-btn to="/todos" class="blue--text" flat>Create One</v-btn>
         </v-card-actions>
       </v-card>
+      <v-dialog
+          persistent
+          v-model="showDeleteConfirm"
+          lazy
+          full-width
+      >
+        <v-card>
+          <v-card-title><span class="headline mb-0">Are you sure?</span></v-card-title>
+          <v-card-text>Are you sure you want to delete these goals(s)? This action is irreversible.</v-card-text>
+          <v-card-actions>
+            <v-btn @click.native.stop="confirmCallback" flat class="blue--text">Yes</v-btn>
+            <v-btn @click.native.stop="abortCallback" flat class="red--text">No</v-btn>
+          </v-card-actions>
+        </v-card>
+      </v-dialog>
     </v-flex>
   </v-layout>
 </template>
@@ -60,7 +75,8 @@ export default {
   },
   data() {
     return {
-      goals: []
+      goals: [],
+      showDeleteConfirm: false
     }
   },
   methods: {
@@ -68,8 +84,21 @@ export default {
       this.$router.push({name: 'goal', params: {id: id}})
     },
     deleteGoal(goal) {
-      this.goals = this.goals.filter((g) => g != goal)
-      this.$root.$data.database.deleteGoal(goal)
+      this.showDeleteConfirm = true
+      this.confirmCallback = () => {
+        this.goals = this.goals.filter((g) => g != goal)
+        goal.todo_ids.forEach((todoId) => {
+          var todo = this.$root.$data.database.getToDo(todoId)
+          if (todo.goalId && todo.goalId == goal._id) {
+            this.$root.$data.database.deleteToDo(todo)
+          }
+        })
+        this.$root.$data.database.deleteGoal(goal)
+        this.showDeleteConfirm = false
+      }
+      this.abortCallback = () => {
+        this.showDeleteConfirm = false
+      }
     }
   },
   components: {
